removing adhoc file. metricFindQuery can be a CHQuery or string
