+++ src/data/CHDatasource.test.ts
@@ -2,9 +2,9 @@ import { toDataFrame } from '@grafana/data';
 import { of } from 'rxjs';
 import { mockDatasource } from '__mocks__/datasource';
 import { CHQuery } from 'types';
+import { AdHocVariableFilter } from './CHDatasource';
 
 interface InstanceConfig {
-  adHocQuery?: string;
   queryResponse: {} | [];
 }
 
@@ -14,7 +14,7 @@ jest.mock('@grafana/runtime', () => ({
   getTemplateSrv: () => templateSrvMock,
 }));
 
-const createInstance = ({ adHocQuery, queryResponse }: Partial<InstanceConfig> = {}) => {
+const createInstance = ({ queryResponse }: Partial<InstanceConfig> = {}) => {
   const instance = mockDatasource;
   jest.spyOn(instance, 'query').mockImplementation((request) => of({ data: [toDataFrame(queryResponse ?? [])] }));
   return instance;
@@ -55,6 +55,17 @@ describe('ClickHouseDatasource', () => {
       const val = await createInstance({}).applyTemplateVariables(query, {});
       expect(spyOnReplace).toHaveBeenCalled();
       expect(val).toEqual({ rawSql });
+    });
+  });
+
+  describe('AdHocManager', () => {
+    it('apply ad hoc filter when there is a where clause', () => {
+      let val = createInstance({}).applyAdHocFilter('SELECT stuff FROM table WHERE col = test;', [{key: 'key', operator: '=', value: 'val'}, {key: 'keyNum', operator: '=', value: '123'}] as AdHocVariableFilter[])
+      expect(val).toEqual(`SELECT stuff FROM table WHERE key = 'val' AND keyNum = 123 AND col = test;`);
+    });
+    it('does not apply ad hoc filter when there is no where clause', () => {
+      let val = createInstance({}).applyAdHocFilter('select stuff from table;', [{key: 'key', operator: '=', value: 'val'}] as AdHocVariableFilter[])
+      expect(val).toEqual('select stuff from table;');
     });
   });
 });
